update to gtk-sharp from cvs.
2004-05-14  Todd Berman  <tberman@sevenl.net>

        * src/ExportCommand.cs:
        * src/PhotoVersionCommands.cs:
        * src/TagCommands.cs: update to gtk-sharp from cvs.

+++ src/ExportCommand.cs
@@ -107,7 +107,7 @@ public class ExportCommand {
 			Glade.XML xml = new Glade.XML (null, "f-spot.glade", "export_gallery_dialog", null);
 			xml.Autoconnect (this);
 
-			export_gallery_dialog.DefaultResponse = (int) ResponseType.Ok;
+			export_gallery_dialog.DefaultResponse = ResponseType.Ok;
 
 			Update ();
